added unsigned keyword in the print

+++ 0x00-hello_world/6-size.c
@@ -7,19 +7,19 @@ int main(void)
 {
 	char character;
 	
-	printf("Size of a char: %c byte(s)\n", sizeof(character));
+	printf("Size of a char: %c byte(s)\n", (unsigned)sizeof(character));
 	int integer;
 	
-	printf("Size of an int: %d byte(s)\n", sizeof(integer));
+	printf("Size of an int: %d byte(s)\n", (unsigned)sizeof(integer));
 	long long_int;
 	
-	printf("Size of a long int: %ld byte(s)\n", sizeof(long_int));
+	printf("Size of a long int: %ld byte(s)\n", (unsigned)sizeof(long_int));
 	long long longer_int;
 	
-	printf("Size of a long long int: %lld byte(s)\n", sizeof(longer_int));
+	printf("Size of a long long int: %lld byte(s)\n", (unsigned)sizeof(longer_int));
 	float decimal;
         
-	printf("Size of a float: %f byte(s)\n", sizeof(decimal));
+	printf("Size of a float: %f byte(s)\n", (unsigned)sizeof(decimal));
 	return (0)
 
 }
